chore(docs | db | config): add docs for docker use | add 3 db's [mysql, postgres, mongodb] | re-arrange folders | connec to docker
