+++ .env.local
@@ -1,3 +1,8 @@
+# App Config
 PORT=5000
-MONGO_URL=mongodb://localhost:27017/sessionbasedauth
 NODE_ENV=development
+
+# Database Connections
+MYSQL_CONN_STRING=mysql://admin:secret@localhost:3306/mydb
+POSTGRES_CONN_STRING=postgresql://admin:secret@localhost:5432/mydb
+MONGODB_CONN_STRING=mongodb://admin:secret@localhost:27017/mydb?authSource=admin
