--- .env.prod
@@ -1,2 +1,0 @@
-PORT=5000
-MONGO_URL=TEST
