+++ .env.sample
@@ -1,6 +1,20 @@
-DATABASE_URL=postgresql://user:password@localhost:5432/yourdb
-PORT=3000
+# Application
+PORT=5000
 NODE_ENV=development
+
+# MySQL
+MYSQL_CONN_STRING=mysql://<username>:<password>@localhost:3306/<database>
+
+# PostgreSQL
+POSTGRES_CONN_STRING=postgresql://<username>:<password>@localhost:5432/<database>
+
+# MongoDB
+MONGODB_CONN_STRING=mongodb://<username>:<password>@localhost:27017/<database>?authSource=admin
+
+
+
 SESSION_COOKIE_NAME=session_id
 SESSION_EXPIRES_IN_MINUTES=30
+
+
