+++ .gitignore
@@ -2,3 +2,7 @@ node_modules
 
 .env.local
 .env.prod
+
+mongo-data
+mysql-data 
+postgres-data
