+++ package.json
@@ -17,12 +17,16 @@
     "cors": "^2.8.5",
     "dotenv": "^16.5.0",
     "express": "^5.1.0",
-    "mongoose": "^8.14.3"
+    "mongodb": "^6.17.0",
+    "mysql2": "^3.14.2",
+    "pg": "^8.16.3"
   },
   "devDependencies": {
     "@types/cors": "^2.8.18",
     "@types/express": "^5.0.1",
-    "@types/node": "^22.15.18",
+    "@types/mysql": "^2.15.27",
+    "@types/node": "^22.16.3",
+    "@types/pg": "^8.15.4",
     "tsx": "^4.19.4",
     "typescript": "^5.8.3"
   }
