+++ src/global.d.ts
@@ -1,3 +1,4 @@
+declare module "*";
 declare global {
   namespace Express {
     interface CustomError extends Error {
